fix: Remove conflicting storage.py file causing uvx package errors
The presence of both storage.py (file) and storage/ (directory) was causing
Python package resolution issues in the uvx environment, resulting in:
  'mcp_memory_service' is not a package

Root cause:
- storage.py tried to import from .storage (circular import)
- Python couldn't resolve which 'storage' to use (file vs directory)

Resolution:
- Removed src/mcp_memory_service/web/oauth/storage.py
- Moved get_oauth_storage() and oauth_storage global to storage/__init__.py
- This makes storage/ the single source of truth for OAuth storage

Fixes 38 failed tests in uvx compatibility CI that were failing with:
  ModuleNotFoundError: No module named 'mcp_memory_service.backup'
